src/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.2.10
exceptiongroup==1.2.2
idna==3.4
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest-django==4.9.0
python-environ==0.4.54
pytz==2023.3.post1
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.4
tomli==2.0.2
typing_extensions==4.10.0
urllib3==1.26.14
